arc.py: semi generic arc builder
takes 3 major pieces of input:
	source x,y,radius
	target x,y,radius
	switches to set some parameters on how many arcs to build, etc

It will look at the source and target regions, find a list of planets
that can build arcs in the source area and a list of unclaimed planets
in the target area. It'll then launch one arc per target planet in a
semi-sorted way (closest arcs first).

Future additions:
	-allow multi arc fleets
	-add ability to attach escort fleet (one frigate i believe would
		keep you from being pirated
	-make the distance sorting a bit smarter
	-set lower boundary on materials remaining on source planets
		ie, dont run a planet into bankruptcy