[basicupgrades] adjust minimum pop for upgrades
In the new DG, the base cost of most of the upgrades increased a bunch.

Based on a quick set of calculations off a spreadsheet, bump the minimum build
order of most of the upgrades to at least let the planet stay afloat at 7%.

TODO: add logic to adjust the tax rate and dynamically adjust.